where do i start
i did a lot of stuff lol. so im using this so you dont get so lost.
thats what this is for right? i guess first off i should explain the new
class, players. its an array of hands. pretty simple i think. its going
to serve as the keylistener. next, in ERS, i took out the parts that
deal with creating hands and replaced them with parts to create a
players object. in hand, i changed the mill function slightly. it does
the same thing but i needed it to use the pile rather than having the
pile use it like it was written to do originally. i also added a slap
function that is dependant on that slap check function i asked you to
write the body of in the pile class. speaking of pile changes, in
addition to adding that spot for the function, i made the arraylist a
static variable. not sure if that was the right thing to do but i think
itll work how i want it to. my idea was if its static i could call
another object of its type and change it to get all of the changes in it
i needed and in the players class called an object of type pile to
manipulate and execute that idea i just mentioned. i put comments next
to all the questionable things i did and in that function i wanted you
to write. sorry if this description doesnt make sense at parts. i was
pretty drained when i wrote it. im hoping i didnt leave out any
important information. i dont think i did. im about to see if i can
still remember everything i need for my test in the morning then going
to bed. night